プレイヤーと敵などのアセット追加
プレイヤー、敵、アイテム、爆弾のアセットを入れた
プレイヤーは仮として入れたため、変更する可能性がある

+++ Assets/Scripts/Player/IApplicableDamage.cs
@@ -2,9 +2,10 @@ using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 
+//ダメージを受けるためのインターフェース
 public interface IApplicableDamage
 {
-    void RecieveDamage(int damage);
+    void RecieveDamage(float damage);
 }
 public interface EnemyDamage
 {
